+++ lib/Terms.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace JanHerman\SimpleTaxonomies;
+
+use Kirby\Cms\Structure;
+
+class Terms extends Structure
+{
+    public const ITEM_CLASS = Term::class;
+}